pyproject.toml
[project]
name = "transit-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tommy Doerr", email = "tommy.b.doerr@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "astral>=3.2",
    "bdfparser>=2.2.0",
    "esptool>=5.2.0",
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pyobjc-framework-quartz>=12.1; sys_platform == 'darwin'",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "rich>=14.3.3",
    "rumps>=0.4.0; sys_platform == 'darwin'",
    "websockets>=16.0",
]

[project.scripts]
transit-tracker = "transit_tracker.cli:main"
transit-tracker-gui = "transit_tracker.gui:main"
transit-tracker-capture = "transit_tracker.capture:main"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0",
    "ruff>=0.9.9",
]

[tool.ruff]
# Target Python 3.13 as specified in [project]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
# Select rules: E (Error), F (Pyflakes), I (isort), B (flake8-bugbear)
select = ["E", "F", "I", "B"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# web.py contains HTML/JS/CSS template strings where line length is impractical
"src/transit_tracker/web.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
markers = [
    "unit: fast isolated tests with mocked dependencies",
    "integration: tests module interactions with local resources",
    "contract: protocol and data compatibility with reference project",
    "e2e: full system tests requiring Docker or live endpoints",
    "docker: marks tests that require Docker (deselect with '-m \"not docker\"')",
]

[tool.coverage.run]
source = ["src/transit_tracker"]
omit = [
    "src/transit_tracker/gui.py",
    "src/transit_tracker/simulator.py",
    "src/transit_tracker/tui.py",
    "src/transit_tracker/monitor.py",
]

[tool.coverage.report]
fail_under = 60
show_missing = true
